# openapi.yaml
openapi: 3.1.0
info:
  title: MinusPod API
  description: |
    REST API for managing podcast feeds with automatic ad detection and removal.


    This API allows you to:
    - Manage podcast feed subscriptions
    - View and manage processed episodes
    - Configure ad detection settings including AI model selection
    - Monitor system status and trigger cleanup operations
    - Manage cross-episode ad patterns with network and podcast scope
    - Submit corrections to improve ad detection accuracy
  version: 2.0.0
  contact:
    name: MinusPod
  license:
    name: MIT
servers:
  # Same-origin relative URL. Swagger UI served from /api/v1/docs uses
  # whatever host the browser is on, so "Try it out" always hits the
  # correct server: localhost in dev, the production domain in prod,
  # a Tailscale/Cloudflare URL on a tunnel, etc. No hardcoded host.
  - url: /api/v1
    description: Same-origin (current server)
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: '{baseUrl}/api/v1'
    description: Custom server
    variables:
      baseUrl:
        default: http://localhost:8000
        description: Base URL of the podcast server
...
paths:
...
  /sponsors:
    get:
      tags:
        - Sponsors
      summary: List all sponsors
      description: Returns list of all known sponsors with their aliases and metadata
      operationId: listSponsors
      parameters:
        - name: include_inactive
          in: query
          description: Include inactive (deleted) sponsors
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of sponsors
          content:
            application/json:
              schema:
                type: object
                properties:
                  sponsors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Sponsor'
    post:
      tags:
        - Sponsors
      summary: Add a new sponsor
      description: Create a new sponsor entry with optional aliases and category
      operationId: addSponsor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Canonical sponsor name
                aliases:
                  type: array
                  items:
                    type: string
                  description: Alternative names for this sponsor
                category:
                  type: string
                  description: Sponsor category
      responses:
        '201':
          description: Sponsor created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Sponsor already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sponsors/{sponsor_id}:
    get:
      tags:
        - Sponsors
      summary: Get a single sponsor
      description: Returns details for a specific sponsor by ID
      operationId: getSponsor
      parameters:
        - $ref: '#/components/parameters/sponsor_id'
      responses:
        '200':
          description: Sponsor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsor'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Sponsors
      summary: Update a sponsor
      description: Update sponsor properties like name, aliases, or category
      operationId: updateSponsor
      parameters:
        - $ref: '#/components/parameters/sponsor_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                aliases:
                  type: array
                  items:
                    type: string
                category:
                  type: string
      responses:
        '200':
          description: Sponsor updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Sponsors
      summary: Delete a sponsor
      description: Deactivate a sponsor (soft delete)
      operationId: deleteSponsor
      parameters:
        - $ref: '#/components/parameters/sponsor_id'
      responses:
        '200':
          description: Sponsor deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /sponsors/normalizations:
    get:
      tags:
        - Sponsors
      summary: List sponsor normalizations
      description: Returns list of sponsor name normalization rules with optional filtering
      operationId: listNormalizations
      parameters:
        - name: category
          in: query
          description: Filter by normalization category
          schema:
            type: string
            enum: [sponsor, url, number, phrase]
        - name: include_inactive
          in: query
          description: Include inactive normalizations
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of normalizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  normalizations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Normalization'
    post:
      tags:
        - Sponsors
      summary: Add a normalization rule
      description: Create a new sponsor name normalization rule using regex pattern matching
      operationId: addNormalization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pattern
                - replacement
                - category
              properties:
                pattern:
                  type: string
                  description: Regex pattern to match against sponsor names
                replacement:
                  type: string
                  description: Canonical replacement name
                category:
                  type: string
                  enum: [sponsor, url, number, phrase]
                  description: Normalization category
      responses:
        '201':
          description: Normalization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /sponsors/normalizations/{norm_id}:
    put:
      tags:
        - Sponsors
      summary: Update a normalization rule
      description: Update an existing normalization rule's pattern, replacement, or category
      operationId: updateNormalization
      parameters:
        - $ref: '#/components/parameters/norm_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pattern:
                  type: string
                  description: Regex pattern to match
                replacement:
                  type: string
                  description: Canonical replacement name
                category:
                  type: string
                  enum: [sponsor, url, number, phrase]
      responses:
        '200':
          description: Normalization updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Sponsors
      summary: Delete a normalization rule
      description: Deactivate a normalization rule (soft delete)
      operationId: deleteNormalization
      parameters:
        - $ref: '#/components/parameters/norm_id'
      responses:
        '200':
          description: Normalization deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Flask session cookie issued by ``POST /auth/login``. Every mutating
        endpoint additionally requires a matching ``X-CSRF-Token`` header
        (double-submit). Public endpoints (``/auth/login``, ``/auth/logout``,
        ``/auth/status``, ``/health``, ``/health/live``) declare ``security: []``.
  parameters:
...
    sponsor_id:
      name: sponsor_id
      in: path
      required: true
      description: Sponsor identifier (integer)
      schema:
        type: integer
        example: 1
    norm_id:
      name: norm_id
      in: path
      required: true
      description: Normalization rule identifier (integer)
      schema:
        type: integer
        example: 1
...
  schemas:
...
    Sponsor:
      type: object
      properties:
        id:
          type: integer
          description: Sponsor ID
        name:
          type: string
          description: Canonical sponsor name
        aliases:
          type: array
          items:
            type: string
          description: Alternative names for this sponsor
        category:
          type:
            - string
            - 'null'
          description: Sponsor category
        common_ctas:
          type: array
          items:
            type: string
          description: Common calls to action associated with this sponsor
        is_active:
          type: boolean
          description: Whether the sponsor is active
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Normalization:
      type: object
      properties:
        id:
          type: integer
          description: Normalization rule ID
        pattern:
          type: string
          description: Regex pattern to match
        replacement:
          type: string
          description: Canonical replacement string
        category:
          type: string
          enum: [sponsor, url, number, phrase]
          description: Normalization category
        is_active:
          type: boolean
          description: Whether the rule is active
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...